+++ src/acconeer/exptool/a121/algo/distance/__main__.py
@@ -23,7 +23,7 @@ detector = Detector(
     detector_config=detector_config,
 )
 
-detector.calibrate_close_range()
+detector.calibrate_detector()
 detector.start()
 
 interrupt_handler = et.utils.ExampleInterruptHandler()